\name{kmrskplot-package}
\alias{kmrskplot-package}
\alias{kmrskplot}
\docType{package}
\title{
\packageTitle{kmrskplot}
}
\description{
\packageDescription{kmrskplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kmrskplot}
\packageIndices{kmrskplot}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{kmrskplot}

Maintainer: \packageMaintainer{kmrskplot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
